# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hive",
    "version": "0.0.1",
    "description": "Hive API"
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/videos": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "Get a list of videos",
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            },
            "in": "query",
            "name": "page",
            "required": true
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "unknown"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "video"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "short"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "stream"
                    ]
                  }
                ]
              }
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "unknown"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "none"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "live"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "upcoming"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "past"
                    ]
                  }
                ]
              }
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "ignored"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "pending"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "done"
                    ]
                  }
                ]
              }
            },
            "in": "query",
            "name": "downloadStatus",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "channelId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "type": "boolean"
            },
            "in": "query",
            "name": "inProgress",
            "required": false
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "videos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "channelId": {
                            "type": "string"
                          },
                          "playlistId": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "string"
                              }
                            ]
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "duration": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "number"
                              }
                            ]
                          },
                          "watchProgress": {
                            "type": "number"
                          },
                          "watchCompleted": {
                            "type": "boolean"
                          },
                          "fileSize": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "number"
                              }
                            ]
                          },
                          "uploadDate": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "number"
                              }
                            ]
                          },
                          "type": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "unknown"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "video"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "short"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "stream"
                                ]
                              }
                            ]
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "unknown"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "none"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "live"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "upcoming"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "past"
                                ]
                              }
                            ]
                          },
                          "downloadStatus": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "ignored"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "pending"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "done"
                                ]
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "channelId",
                          "playlistId",
                          "title",
                          "description",
                          "duration",
                          "watchProgress",
                          "watchCompleted",
                          "fileSize",
                          "uploadDate",
                          "type",
                          "status",
                          "downloadStatus"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "videos",
                    "total"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/videos/ignore": {
      "post": {
        "tags": [
          "Videos"
        ],
        "description": "Ignore videos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "videoIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "videoIds"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "204": {
            "description": "Videos IDs ignored"
          }
        }
      }
    },
    "/videos/scan": {
      "post": {
        "tags": [
          "Videos"
        ],
        "description": "Scan all channels for new videos",
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "201": {
            "description": "Scan started successfully"
          }
        }
      }
    },
    "/videos/{videoId}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "Get a video",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "videoId",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "channelId": {
                      "type": "string"
                    },
                    "playlistId": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "duration": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "watchProgress": {
                      "type": "number"
                    },
                    "watchCompleted": {
                      "type": "boolean"
                    },
                    "fileSize": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "uploadDate": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "video"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "short"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "stream"
                          ]
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "none"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "live"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "upcoming"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "past"
                          ]
                        }
                      ]
                    },
                    "downloadStatus": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ignored"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pending"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "channelId",
                    "playlistId",
                    "title",
                    "description",
                    "duration",
                    "watchProgress",
                    "watchCompleted",
                    "fileSize",
                    "uploadDate",
                    "type",
                    "status",
                    "downloadStatus"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Video not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Videos"
        ],
        "description": "Update a video",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "watchProgress": {
                    "type": "number"
                  },
                  "watchCompleted": {
                    "type": "boolean"
                  },
                  "downloadStatus": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "ignored"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "pending"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "videoId",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "channelId": {
                      "type": "string"
                    },
                    "playlistId": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "duration": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "watchProgress": {
                      "type": "number"
                    },
                    "watchCompleted": {
                      "type": "boolean"
                    },
                    "fileSize": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "uploadDate": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "video"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "short"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "stream"
                          ]
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "none"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "live"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "upcoming"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "past"
                          ]
                        }
                      ]
                    },
                    "downloadStatus": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ignored"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pending"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "channelId",
                    "playlistId",
                    "title",
                    "description",
                    "duration",
                    "watchProgress",
                    "watchCompleted",
                    "fileSize",
                    "uploadDate",
                    "type",
                    "status",
                    "downloadStatus"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}